working on both contact and user route

+++ controllers/contactController.js
@@ -2,7 +2,7 @@ const Contacts = require("../models/contacts.model");
 
 //@desc Create New contact
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = async (req, res) => {
   try {
     const { name, email, phone } = req.body;
@@ -10,6 +10,7 @@ const createContact = async (req, res) => {
       return res.status(400).send({ massgae: "All field are mendatory " });
     }
     const newContact = new Contacts({
+      user_id: req.user.userId,
       name,
       email,
       phone,
@@ -24,11 +25,12 @@ const createContact = async (req, res) => {
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 
 const getAllContacts = async (req, res) => {
   try {
-    const contacts = await Contacts.find();
+    console.log(req.user.userId);
+    const contacts = await Contacts.find({ user_id: req.user.userId });
     res.status(200).json(contacts);
   } catch (error) {
     res.status(500).send(error.message);
@@ -37,7 +39,7 @@ const getAllContacts = async (req, res) => {
 
 //@desc Get contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = async (req, res) => {
   try {
     const contact = await Contacts.findById(req.params.id);
@@ -52,13 +54,20 @@ const getContact = async (req, res) => {
 
 //@desc Update contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = async (req, res) => {
   try {
     const contact = await Contacts.findById(req.params.id);
     if (!contact) {
       return res.status(404).json({ massgae: "Contact not found" });
     }
+
+    if (contact.user_id.toString() !== req.user.id) {
+      return res
+        .status(403)
+        .json("User don't have permission to update other user contacts");
+    }
+
     const updateContact = await Contacts.findByIdAndUpdate(
       req.params.id,
       req.body,
@@ -75,13 +84,19 @@ const updateContact = async (req, res) => {
 
 //@desc Delete contact
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = async (req, res) => {
   try {
     const contact = await Contacts.findById(req.params.id);
     if (!contact) {
       return res.status(404).json({ massgae: "Contact not found" });
     }
+    if (contact.user_id.toString() !== req.user.id) {
+      return res
+        .status(403)
+        .json("User don't have permission to update other user contacts");
+    }
+    // await Contact.deleteOne({ _id: req.params.id });
     await Contacts.findByIdAndRemove(req.params.id);
     res.status(200).json({ massgae: `delete contact for ${req.params.id}` });
   } catch (error) {
